Move RolePrincipal to role package
