+++ controller-server/src/main/java/com/yahoo/vespa/hosted/controller/restapi/filter/Auth0RoleFilter.java
@@ -3,6 +3,7 @@ package com.yahoo.vespa.hosted.controller.restapi.filter;
 import com.yahoo.jdisc.http.filter.DiscFilterRequest;
 import com.yahoo.jdisc.http.filter.security.base.JsonSecurityRequestFilterBase;
 import com.yahoo.vespa.hosted.controller.role.RoleMembership;
+import com.yahoo.vespa.hosted.controller.role.RolePrincipal;
 
 import java.util.Objects;
 import java.util.Optional;
